backend/services/ai-classifier-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
transformers==4.35.2
pika==1.3.2
python-dotenv==1.0.0
